requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.6.0
pydantic==2.10.0
pydantic[email]
python-multipart==0.0.6
jinja2==3.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
pillow>=10.0.0
bcrypt==4.2.0
